pyproject.toml
[project]
name = "multiagentlaptoprecommendation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-adk>=1.19.0",
    "python-dotenv>=1.2.1",
]
